requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
Pillow==11.3.0
django-filter==25.1
django-cors-headers==4.7.0
gunicorn==23.0.0
django-storages==1.14.6
boto3==1.39.6
celery==5.5.3
redis==6.2.0
djangorestframework-simplejwt==5.5.0
pandas==2.3.1
sqlalchemy==2.0.41
django-celery-results==2.6.0